+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 5.7.4
+----------------
+- Increase Currency input max length to 8 in Configuration.php
+
 Changes in 5.7.3
 ----------------
 - Fix wkhtmltopdf exec crash: Old code for Windows OS in Wkhtmltopdf.php